Seperate out classes from parser.py into their own files to better arrange the code. Import the main classes into __init__.py
