+++ app_pack_generator/parser.py
@@ -1,15 +1,13 @@
 import os
 import copy
 import yaml
-import requests
-import datetime
-import subprocess
 
-import git
 import json
-import docker
 import papermill
 import jsonschema
+
+from .util import Util
+from .git import GitHelper
 
 """
 Loads the .ipynb jsonschema for validation purposes (TBD).
@@ -20,240 +18,6 @@ SCHEMA_LIST = [os.path.join(LOCAL_PATH,
                             'schemas/nbformat.v4.{v}.schema.json'.format(v=i)) for i in range(0, 6)]
 INPUT_TAG = 'parameters'
 OUTPUT_TAG = 'outputFiles'
-REPO2DOCKER_ENV = os.getenv('env')
-
-
-class Util:
-    @staticmethod
-    def System(cmd):
-        """Runs a terminal commands with the specified command."""
-        return subprocess.run(cmd,
-                              stdout=subprocess.PIPE, stderr=subprocess.PIPE, universal_newlines=True)
-
-    @staticmethod
-    def TimeFunction(func, *args, **kwargs):
-        """
-        Times the execution of a function and is arguments.
-
-        As this function does not utilize a decorator, it can
-        be used to time any arbitrary function call as opposed
-        to only the ones the programmer declares.
-        """
-        start = datetime.datetime.now()
-        ret = func(*args, **kwargs)
-        time_diff = datetime.datetime.now() - start
-        return time_diff.total_seconds() * 1000, ret
-
-    @staticmethod
-    def GetKeyType(inferred_type, default_val):
-        """Attempts to convert the inferred type to an equivalent CWL type.
-
-        Otherwise, checks if a given key's default value is a string or can be converted to a float."""
-        inferred_type = inferred_type.lower()
-        convert_dict = {
-            'string': ['stage_in', 'stage-in', 'string'],
-            'File': ['stage_out', 'stage-out', 'file'],
-            'int': ['int', 'integer'],
-            'boolean': ['bool', 'boolean'],
-            'float': ['float'],
-            'double': ['double'],
-            'Directory': ['directory'],
-            'Any': ['any']
-        }
-        for key in convert_dict:
-            if inferred_type in convert_dict[key]:
-                return key
-
-        if default_val.find('"') != -1 or default_val.find('\'') != -1:
-            return 'string'
-        key_type = 'Any'
-        try:
-            temp = float(default_val)
-            key_type = 'float'
-        except ValueError:
-            pass
-        return key_type
-
-    @staticmethod
-    def WriteYMLFile(fname, target):
-        """Writes [target] dictionary to a .yml file with filename [fname]."""
-        with open(fname, 'w', encoding='utf-8') as f:
-            f.write("#!/usr/bin/env cwl-runner\n")
-            yaml.dump(target, f, default_flow_style=False)
-
-    @staticmethod
-    def DownloadLink(url, default=None):
-        """Downloads the specified URL via a GET request.
-
-        If url is not a valid link, or if the request fails, returns the [default]
-        parameter instead.
-        """
-        try:
-            response = requests.get(url)
-            if response.status_code == 404:
-                raise RuntimeError('<Response 404>')
-            return response
-        except Exception as e:
-            if url.startswith('http://') or url.startswith('https://'):
-                print('Could not download assumed URL: \'' + url + '\'')
-                print(e)
-        return default
-
-
-class GitHelper:
-    def __init__(self, url, dst=os.getcwd()):
-        """Manages a freshly cloned git repository."""
-        self.repo = GitHelper.Clone(url, dst)
-        self.url = url
-        self.directory = dst
-
-        split = url.replace('.git', '').split('/')
-        self.owner = split[-2].strip()
-        self.name = split[-1].strip()
-        self.checkout = 'HEAD'
-        self.dirname = self.owner + '/' + self.name + '/' + self.checkout
-
-    def Checkout(self, arg):
-        """Runs the checkout command on this repository.
-
-        'arg' is either a commit hash, a tag, or a branch name.
-        Initializes any new submodules as well.
-        """
-        self.repo.git.checkout(arg)
-        self.repo.git.submodule('update', '--init')
-
-        self.checkout = arg
-        self.dirname = self.owner + '/' + self.name + '/' + self.checkout
-
-    @staticmethod
-    def Clone(repolink, dst=os.getcwd()):
-        """Clones the specified repository using its HTTPS URL."""
-        print('Cloning to ' + dst + '...')
-        return git.Repo.clone_from(repolink, dst)
-
-    @staticmethod
-    def Message(repodir):
-        """Return the commit message for the current commit."""
-        # git.Repo(localdir)
-        process = Util.System(['git', 'log', '-1', '--pretty=%B'])
-        if process.stdout != '':
-            return process.stdout.strip()
-        return process.stderr.strip()
-
-    @staticmethod
-    def CommitHash(repodir):
-        """Return the commit message for the current commit."""
-        # git.Repo(localdir)
-        process = Util.System(['git', 'rev-parse', 'HEAD'])
-        if process.stdout != '':
-            return process.stdout.strip()
-        return process.stderr.strip()
-
-    @staticmethod
-    def RemoteURL(repodir, https=True):
-        """Return the remote URL for the current repository."""
-        process = Util.System(['git', 'config', '--get', 'remote.origin.url'])
-        if process.stdout != '':
-            if https:
-                process.stdout = process.stdout.strip()
-                process.stdout = process.stdout.replace(':', '/')
-                process.stdout = process.stdout.replace('git@', 'https://')
-            return process.stdout.strip()
-        return process.stderr.strip()
-
-    @staticmethod
-    def Push(repodir):
-        """Pushes all changes from the specified repository."""
-        repo = git.Repo(repodir)
-        repo.git.add(update=True)
-        repo.index.commit('Automatic push from Jenkins.')
-        origin = repo.remote(name='origin')
-        origin.push()
-
-    @staticmethod
-    def GetTag(repodir):
-        """Gets the current repository tag associated with the current commit."""
-        process = Util.System(['git', 'describe', '--tags', '--abbrev=0'])
-        if process.stdout != '':
-            return process.stdout.strip()
-        return process.stderr.strip()
-
-
-class DockerUtil:
-
-    def __init__(self, repo, do_prune=True):
-        self.repo = repo
-        self.do_prune = do_prune
-
-        self.docker_client = docker.from_env()
-        username = os.getenv('DOCKER_USER')
-        password = os.getenv('DOCKER_PASS')
-        self.docker_client.login(username=username, password=password)
-
-    def Repo2Docker(self):
-        """Calls repo2docker on the local git directory to generate the Docker image.
-
-        No further modifications are made to the docker image.
-        """
-
-        # Prune all dangling containers and images to reclaim space and prevent cache usage.
-        if self.do_prune:
-            try:
-                self.docker_client.containers.prune()
-                self.docker_client.images.prune()
-            except requests.exceptions.ReadTimeout as e:
-                print('An error occurred while pruning: {}'.format(e.what()))
-
-        # Repo2Docker call using the command line.
-        image_tag = self.repo.owner + '.' + self.repo.name + '.' + self.repo.checkout
-        if len(image_tag) > 128:
-            image_tag = image_tag[0:128]
-
-        # Clean up image_tag to confirm to Docker rules
-        # 1. Make sure all characters are lower case
-        # 2. Remove repeated periods (ie ..) in cases where repo.name is empty
-        image_tag = image_tag.lower()
-        image_tag = image_tag.replace('..', '.')
-
-        cmd = ['jupyter-repo2docker', '--user-id', '1000', '--user-name', 'jovyan',
-               '--no-run', '--debug', '--image-name', image_tag, self.repo.directory]
-        if REPO2DOCKER_ENV is not None and os.path.exists(REPO2DOCKER_ENV):
-            cmd = ['jupyter-repo2docker', '--user-id', '1000', '--user-name', 'jovyan',
-                   '--no-run', '--debug', '--image-name', '--config',
-                   REPO2DOCKER_ENV, image_tag, self.repo.directory]
-        elif REPO2DOCKER_ENV is not None:
-            response = Util.DownloadLink(REPO2DOCKER_ENV)
-            if response is not None:
-                with open(os.path.join(self.repo.directory, 'Dockerfile'), 'w') as f:
-                    f.write(response.text)
-            else:
-                msg = 'Failed to download the specified configuration file: ' + REPO2DOCKER_ENV
-                raise RuntimeError(msg)
-
-        process = Util.System(cmd)
-        print(process.stdout)
-        print(process.stderr)
-
-        return image_tag
-
-    def PushImage(self, image_tag, registry):
-
-        # Save the newly created image to a tarball if the build succeeded.
-        image = self.docker_client.images.get(image_tag)
-        image.tag(registry, tag=image_tag)
-        for line in self.docker_client.api.push(registry, tag=image_tag, stream=True, decode=True):
-            print(line)
-
-        if self.do_prune:
-            self.docker_client.images.remove(image.id, force=True)
-            try:
-                self.docker_client.containers.prune()
-                self.docker_client.images.prune()
-            except requests.exceptions.ReadTimeout as e:
-                print('An error occurred while pruning: {}'.format(e.what()))
-
-        return registry + ':' + image_tag
-
 
 class AppNB:
     """Defines a parsed Jupyter Notebook read as a JSON file."""
